Save and continue with defaults, remove disabled attr on button, show errors instead
